backend/requirements.txt
django==4.2.13
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
tzlocal==5.2
gunicorn==22.0.0
psycopg2==2.9.9
flake8==7.0.0